+++ 3rdparty/untz/src/UserAudioSource.h
@@ -8,7 +8,7 @@ class UserAudioSource : public AudioSource
 public:
 	UserAudioSource(UInt32 sampleRate, UInt32 numChannels, UNTZ::StreamCallback* callback, void* userdata);
 	// AudioSource
-	virtual UInt32 readFrames(float* buffer, UInt32 numChannels, UInt32 numSamples);
+	virtual Int64 readFrames(float* buffer, UInt32 numChannels, UInt32 numSamples);
 private:
 	UNTZ::StreamCallback* mCallback;
 	void* mUserdata;
